move user reg components
