+++ src/app/components/index.ts
@@ -1,9 +1,3 @@
-export { default as AcceptTOS } from './AcceptTOS';
-export { default as AgeConfirmation } from './AgeConfirmation';
-export { default as PartnerPlugins } from './PartnerPlugins';
-export { default as RegisterUser } from './RegisterUser';
-export { default as SelectPlan } from './SelectPlan';
-export { default as SpecialOffers } from './SpecialOffers';
 export { default as Success } from './Success';
 export { default as Error } from './Error';
-export { default as ReasonDemo } from './ReasonDemo';
+export { default as ReasonDemo } from './chemli/ReasonDemo';
