+++ src/app/components/registration/AcceptTOS.tsx
