+++ src/app/components/registration/AgeConfirmation.tsx
