+++ src/app/components/registration/PartnerPlugins.tsx
