+++ src/app/components/registration/SelectPlan.tsx
