+++ src/app/components/registration/index.ts
@@ -1,0 +1,6 @@
+export { default as AcceptTOS } from './AcceptTOS';
+export { default as AgeConfirmation } from './AgeConfirmation';
+export { default as PartnerPlugins } from './PartnerPlugins';
+export { default as RegisterUser } from './RegisterUser';
+export { default as SelectPlan } from './SelectPlan';
+export { default as SpecialOffers } from './SpecialOffers';
